Update the global report list immediately before send a request for “Save” or “Delete” to create a more responsive application. After the requests, update the reports with real data.

+++ src/ReportEditor.svelte
@@ -56,6 +56,11 @@
 
   let deleteReport = async e => {
     e.preventDefault();
+    const needDelete = confirm("Delete report?");
+    if (!needDelete) {
+      return;
+    }
+
     await reports.deleteReport(reportInEditMode.id);
   };
